third_party/blink/web_tests/flag-specific/highdpi/css3/calc/table-border-spacing-expected.txt

FAIL document.getElementById("stimulus").offsetWidth should be 416. Was 413.
FAIL document.getElementById("stimulus").offsetHeight should be 205. Was 203.
PASS successfullyParsed is true

TEST COMPLETE

				
				
